requirements.txt
pandas
openpyxl
python-magic-bin
